feat: refactor db
